arch/riscv: ePMP: Fixup ePMP detection
We currently support detecting any number of ePMP regions by writing
values and seeing if the new value sticks. This allows us to support
multiple platforms, but it doesn't work if the bootloader has already
enabled MML.

This patch changes the behaviour if MML has been set, so that we ensure
we won't accidently lock ourselves out.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>