arch/rv32i: ePMP: Add catch all default regions during setup
A previous boot stage has enabled Machine Mode Whitelist Policy
(mseccfg.MMWP), which we can't disable. That means that Tock will
be denied a memory access if it doesn't match a PMP region
This eventually won't matter, as Tock configures it's own regions,
but it makes it difficult to setup.

To allow us to setup the ePMP regions without locking ourselves out
we create two special regions.

We create an allow all region as the first and last region.
This way no matter what the previous stage did, we should have a
working fallback while we modify the rules.
If the previous stage has an allow all in the first or last region,
we won't get locked out. If the previous stage created a range of
regions we also won't get locked out.

Signed-off-by: Alistair Francis <alistair.francis@wdc.com>